Calculate Casualties

Calculate the number of casualties grouped by priogrid_gid and period and side specified when the RebelTrackDataFrameGid was created with rebeltrack_dataframe_gid. Specifically for side B

rebeltrack_casualty_count_gid_sideB(x, var, fill = 0, lag = 0, weight = NULL)

Arguments

x

A RebelTrackDataFrameGid object

var

Name of the variable to create

fill

Default value when no events are observed in the specified period.

lag

An integer giving the number of positions to lead or lag by.

weight

A weight matrix to apply. Must be a Rebeltrack GID matrix to respect dimensionality.

Examples

library(rebeltrack)

if (FALSE) { # \dontrun{
dataset <- rebeltrack_load_dataset(type = UCDP_STATE_BASED)

data <- rebeltrack_dataframe_gid(dataset,
                             period = "month",
                             balanced = TRUE)

data <- rebeltrack_casualty_count_gid_sideB(data, casualty_count)
} # }